Last commit before making repo public
Fixed a bunch of misc. stuff before making this repo public

Even though Biological Anomaly's source code is obsolete and very outdated, I thought it'd still be nice to release it after all these years

The reboot (whenever that happens) will be made on Godot, using parts of the Vs. Purki Christmas source code as a base, which is far more optimized and functional than this old code

Have fun looking through this mess of a code I made when I was like 15 :)